+++ test/dedicated_test_failures/framework.jl
@@ -12,6 +12,7 @@ import EcologicalNetworksDynamics.Framework:
     ConflictWithSystemComponent,
     Framework,
     HookCheckFailure,
+    InconsistentForSameComponent,
     ItemMacroExecError,
     ItemMacroParseError,
     MissingRequiredComponent,